Modified client so it works with Marketo's API.
The default client uses the normal '<soap:>' element, while Marketo's
API uses a custom '<SOAP-ENV:>' element. The client provides no way to
modifiy this so had to hard code it to work with Marketo. This is just a
quick fix to prove that this can work. Next step will be to create a
custom MarketoClient that inherits from Client and overwrites _inoke().